removed all postburner folders and moved everything into reco folders
