+++ src/algorithms/CMakeLists.txt
@@ -8,7 +8,6 @@ add_subdirectory(pid_lut)
 add_subdirectory(digi)
 add_subdirectory(reco)
 add_subdirectory(fardetectors)
-add_subdirectory(postburn)
 
 if(USE_ONNX)
   add_subdirectory(onnx)
