+++ src/factories/CMakeLists.txt
@@ -3,4 +3,3 @@ add_subdirectory(digi)
 add_subdirectory(fardetectors)
 add_subdirectory(tracking)
 add_subdirectory(meta)
-add_subdirectory(postburn)
