--- src/factories/postburn/CMakeLists.txt
@@ -1,2 +1,0 @@
-set(PLUGIN_NAME "factories_postburn")
-plugin_headers_only(${PLUGIN_NAME})
